requirements.txt
python-telegram-bot
openai
python-dotenv
google-genai python-dotenv